I did some godly shit this time
Updated ApiHeartbeatService to now log the heartbeat of all 4 microservices. (They may or may not be working containerized btw)

Fixed appsettings so that the above change actually worked. Tyler Campbell is my goat.

Also added HealthController.cs for Snake, Pong, and Tetris. Idk if those are needed, but they are there now.